Module Github.Milestone

The Milestone module exposes GitHub's milestone API.

val for_repo : ?⁠token:Token.t -> ?⁠state:Filter.state -> ?⁠sort:Filter.milestone_sort -> ?⁠direction:Filter.direction -> user:string -> repo:string -> unit -> Github_t.milestone Stream.t

for_repo ?state ?sort ?direction ~user ~repo () is a stream of all milestones in repo user/repo which match ?state (default `Open). The stream is sorted by ?sort (default `Due_date) and ordered by ?direction (default `Desc).

val get : ?⁠token:Token.t -> user:string -> repo:string -> num:int -> unit -> Github_t.milestone Response.t Monad.t

get ~user ~repo ~num () is milestone number num in repo user/repo.

val create : ?⁠token:Token.t -> user:string -> repo:string -> milestone:Github_t.new_milestone -> unit -> Github_t.milestone Response.t Monad.t

create ~user ~repo ~milestone () is the newly created milestone described by milestone in repo user/repo.

val delete : ?⁠token:Token.t -> user:string -> repo:string -> num:int -> unit -> unit Response.t Monad.t

delete ~user ~repo ~num () activates after milestone num in repo user/repo has been deleted.

val update : ?⁠token:Token.t -> user:string -> repo:string -> milestone:Github_t.update_milestone -> num:int -> unit -> Github_t.milestone Response.t Monad.t

update ~user ~repo ~milestone ~num () is the updated milestone num in repo user/repo as described by milestone.

val labels : ?⁠token:Token.t -> user:string -> repo:string -> num:int -> unit -> Github_t.label Stream.t

labels ~user ~repo ~num () is a stream of all labels for milestone num in repo user/repo.