Module Github.Pull

The Pull module contains functionality relating to GitHub's pull request API.

val for_repo : ?token:Token.t -> ?state:Filter.state -> user:string -> repo:string -> unit -> Github_t.pull Stream.t

for_repo ?state ~user ~repo () is a stream of pull requests against repo user/repo which are currently in state ?state (default `Open).

val get : ?token:Token.t -> user:string -> repo:string -> num:int -> unit -> Github_t.pull Response.t Monad.t

get ~user ~repo ~num () is the pull request user/repo#num.

val create : ?token:Token.t -> user:string -> repo:string -> pull:Github_t.new_pull -> unit -> Github_t.pull Response.t Monad.t

create ~user ~repo ~pull () is the newly created pull request against repo user/repo as described by pull.

val create_from_issue : ?token:Token.t -> user:string -> repo:string -> pull_issue:Github_t.new_pull_issue -> unit -> Github_t.pull Response.t Monad.t

create_from_issue ~user ~repo ~pull_issue () is the newly created pull request from an issue against repo user/repo as described by pull_issue.

val update : ?token:Token.t -> user:string -> repo:string -> update_pull:Github_t.update_pull -> num:int -> unit -> Github_t.pull Response.t Monad.t

update ~user ~repo ~update_pull ~num () is the updated pull request user/repo#num as described by update_pull.

val commits : ?token:Token.t -> user:string -> repo:string -> num:int -> unit -> Github_t.commit Stream.t

commits ~user ~repo ~num () is the stream of commits included in pull request user/repo#num.

val files : ?token:Token.t -> user:string -> repo:string -> num:int -> unit -> Github_t.file Stream.t

files ~user ~repo ~num () is the stream of files included in pull request user/repo#num.

val is_merged : ?token:Token.t -> user:string -> repo:string -> num:int -> unit -> bool Response.t Monad.t

is_merged ~user ~repo ~num () is true if pull request user/repo#num has been merged.

val merge : ?token:Token.t -> user:string -> repo:string -> num:int -> ?merge_commit_message:string -> unit -> Github_t.merge Response.t Monad.t

merge ~user ~repo ~num ?merge_commit_message () is the merge of pull request user/repo#num with optional commit message ?merge_commit_message.