Module Github.Rate_limit

The Rate_limit module contains explicit rate limit API request functions which do not read the rate limit cache but do write to it.

val all : ?⁠token:Token.t -> unit -> Github_t.rate_resources Monad.t

all () is the current token's rate limit information for all rate limiting regimes.

val for_core : ?⁠token:Token.t -> unit -> Github_t.rate Monad.t

for_core () is the current token's rate limit information for the Core rate limit regime.

val for_search : ?⁠token:Token.t -> unit -> Github_t.rate Monad.t

for_search () is the current token's rate limit information for the Search rate limit regime.