Module Github.Search

The Search module exposes GitHub's search interfaces.

val repos : ?⁠token:Token.t -> ?⁠sort:Filter.repo_sort -> ?⁠direction:Filter.direction -> qualifiers:Filter.qualifier list -> keywords:string list -> unit -> Github_t.repository_search Stream.t

repos ?sort ?direction ~qualifiers ~keywords () is a stream of repository search results for keywords and matching qualifiers predicates. Results are sorted by ?sort (default best match) and ordered by ?direction (default `Desc).

val issues : ?⁠token:Token.t -> ?⁠sort:Filter.repo_sort -> ?⁠direction:Filter.direction -> qualifiers:Filter.issue_qualifier list -> keywords:string list -> unit -> Github_t.repository_issue_search Stream.t

issues ?sort ?direction ~qualifiers ~keywords () is a stream of issue search results for keywords and matching qualifiers predicates. Results are sorted by ?sort (default best match) and ordered by ?direction (default `Desc).