Module Github.Status

The Status module provides the functionality of GitHub's status API.

val for_ref : ?⁠token:Token.t -> user:string -> repo:string -> git_ref:string -> unit -> Github_t.status Stream.t

for_sha ~user ~repo ~git_ref () is a stream of statuses attached to the SHA, branch name, or tag name git_ref in repo user/repo.

val create : ?⁠token:Token.t -> user:string -> repo:string -> sha:string -> status:Github_t.new_status -> unit -> Github_t.status Response.t Monad.t

create ~user ~repo ~sha ~status () is a newly created status on SHA sha in repo user/repo as described by status.

val get : ?⁠token:Token.t -> user:string -> repo:string -> sha:string -> unit -> Github_t.combined_status Response.t Monad.t

get ~user ~repo ~sha () is the combined status of the ref sha in the repo user/repo.