Module Github.Git_obj

The Git_obj module contains utility functions for working with git concepts.

val type_to_string : Github_t.obj_type -> string

type_to_string type is the string name of object type type.

val split_ref : string -> string * string

split_ref ref is the pair of ref directory and ref name. E.g. if ref is "refs/tags/foo/bar" then split_ref ref is ("tags","foo/bar").